refactor(examples): rename _base harness folder to _layout
